Extract duplicate cue visual span map functions into one helper
